does not scale parental phenotype
I’ve updated GE so that it gives the correct values of VF under F-F transmission. The problem before is that I was scaling the parental phenotype relative to the parental phenotype in the gen0 (or more technically, the parents of gen0) generation. This was disallowing VF to go where it wanted to go. I’ve checked that F-F transmission now goes down by half when p=m=.5, and similarly for other values of p & m.